Add new shared connection establishment module
This module will handle connection establishment for both SENDRECV and
RDMA transports (they will be ported in the next commit).

Features of new connection manager:
- Supports multiple accept calls per listen communicator
- Supports semantics of NCCL's shared comm/multi-recv connection
  establishment (although this feature is not yet supported by the
  transports)

+++ include/nccl_ofi.h
@@ -221,7 +221,7 @@ static_assert(sizeof(nccl_ofi_connection_info_t) == 80, "Wrong size for SENDRECV
 
 typedef struct nccl_net_ofi_conn_handle {
 	char ep_name[MAX_EP_ADDR];
-	uint32_t comm_id;
+	uint64_t comm_id;
 	/* Save temporary communicator state when creating send communicator */
 	save_comm_state_t state;
 } nccl_net_ofi_conn_handle_t;
